fix circular import on newer branch
