+++ file_service.py
@@ -1,11 +1,14 @@
 import os
-import utils
-from utils import print_location_error
+import random
+import string
+
+FILE_NAME_LENGTH = 10
+LETTERS_DIGITS = string.ascii_letters + string.digits
 
 
 def create_file(content=''):
     try:
-        filename = f'{utils.generate_name()} .txt'
+        filename = f'{generate_name()} .txt'
         with open(filename, 'w') as file:
             file.write(content)
         print(f'File {filename} was created in {os.getcwd()} directory.')
@@ -37,3 +40,11 @@ def get_metadata(filename):
         return metadata
     except FileNotFoundError:
         print_location_error()
+
+
+def print_location_error():
+    print("File not found.\n")
+
+
+def generate_name() -> str:
+    return ''.join(random.choice(LETTERS_DIGITS) for _ in range(FILE_NAME_LENGTH))
