+++ utils.py
@@ -1,13 +1,4 @@
-import random
-import string
 from file_service import create_file, read_file, delete_file, get_metadata
-
-FILE_NAME_LENGTH = 10
-LETTERS_DIGITS = string.ascii_letters + string.digits
-
-
-def generate_name() -> str:
-    return ''.join(random.choice(LETTERS_DIGITS) for _ in range(FILE_NAME_LENGTH))
 
 
 def show_options() -> str:
@@ -37,10 +28,6 @@ class InvalidOption(Exception):
 
 def print_option_error():
     print("Choose anohter option.\n")
-
-
-def print_location_error():
-    print("File not found.\n")
 
 
 def app():
